endpoint/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
pydantic==2.11.7
pydantic-settings==2.9.1
requests==2.32.3
psycopg2-binary==2.9.10
qdrant-client==1.14.2
kafka-python==2.0.2
